removed Payload-structure and added marshal.go
