+++ nist/residue.go
@@ -15,8 +15,22 @@ import (
 
 type residuePoint struct {
 	big.Int
-	g *ResidueGroup
-	abstract.Payload
+	g     *ResidueGroup
+	Suite string
+}
+
+func (rp *residuePoint) GetSuite() abstract.Suite {
+	s, _ := abstract.StringToSuite(rp.Suite)
+	return s
+}
+
+func (rp *residuePoint) SetSuite(s abstract.Suite) {
+	rp.Suite = s.String()
+}
+
+// Point creates a Point-structure from a PointInterface
+func (rp *residuePoint) Point() *abstract.Point {
+	return &abstract.Point{rp}
 }
 
 // Steal value from DSA, which uses recommendation from FIPS 186-3
@@ -35,12 +49,12 @@ func (p *residuePoint) Equal(p2 *abstract.Point) bool {
 
 func (p *residuePoint) Null() *abstract.Point {
 	p.Int.SetInt64(1)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) Base() *abstract.Point {
 	p.Int.Set(p.g.G)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) Valid() bool {
@@ -74,7 +88,7 @@ func (p *residuePoint) Pick(data []byte, rand cipher.Stream) (*abstract.Point, [
 		}
 		p.Int.SetBytes(b)
 		if p.Valid() {
-			return p.MakePoint(p), data[dl:]
+			return p.Point(), data[dl:]
 		}
 	}
 }
@@ -96,27 +110,27 @@ func (p *residuePoint) Data() ([]byte, error) {
 func (p *residuePoint) Add(a, b *abstract.Point) *abstract.Point {
 	p.Int.Mul(&a.PointInterface.(*residuePoint).Int, &b.PointInterface.(*residuePoint).Int)
 	p.Int.Mod(&p.Int, p.g.P)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) Sub(a, b *abstract.Point) *abstract.Point {
 	binv := new(big.Int).ModInverse(&b.PointInterface.(*residuePoint).Int, p.g.P)
 	p.Int.Mul(&a.PointInterface.(*residuePoint).Int, binv)
 	p.Int.Mod(&p.Int, p.g.P)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) Neg(a *abstract.Point) *abstract.Point {
 	p.Int.ModInverse(&a.PointInterface.(*residuePoint).Int, p.g.P)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) Mul(b *abstract.Point, s *abstract.Secret) *abstract.Point {
 	if b == nil {
-		return p.Base().Mul(p.MakePoint(p), s)
+		return p.Base().Mul(p.Point(), s)
 	}
 	p.Int.Exp(&b.PointInterface.(*residuePoint).Int, &s.SecretInterface.(*Int).V, p.g.P)
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 func (p *residuePoint) MarshalSize() int {
@@ -141,12 +155,12 @@ func (p *residuePoint) UnmarshalBinary(data []byte) error {
 }
 
 func (p *residuePoint) MarshalTo(w io.Writer) (int, error) {
-	return group.PointMarshalTo(p.MakePoint(p), w)
+	return group.PointMarshalTo(p.Point(), w)
 }
 
 func (p *residuePoint) UnmarshalFrom(r io.Reader) (int, error) {
 	fmt.Println("Unmarshalfrom residue")
-	return group.PointUnmarshalFrom(p.MakePoint(p), r)
+	return group.PointUnmarshalFrom(p.Point(), r)
 }
 
 /*
@@ -207,7 +221,7 @@ func (g *ResidueGroup) PointLen() int { return (g.P.BitLen() + 7) / 8 }
 func (g *ResidueGroup) Point() *abstract.Point {
 	p := new(residuePoint)
 	p.g = g
-	return p.MakePoint(p)
+	return p.Point()
 }
 
 // Returns the order of this Residue group, namely the prime Q.
